+++ src/app/app-routing.module.ts
@@ -2,13 +2,17 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import {AboutComponent} from './about/about.component';
 import {NotFoundComponent} from './not-found/not-found.component';
+import { LoginComponent } from './login/login.component';
 
 
 const routes: Routes = [
+  {path: 'login', component: LoginComponent},
   {path: 'about', component: AboutComponent},
+  //Page to redirect to on initial load
+  {path: '', redirectTo: 'login', pathMatch: 'full'},
   // Catch all non-existing paths
-  {path: 'about', component: NotFoundComponent},
-  {path: '**', redirectTo: '/not-found', pathMatch: 'full'}
+  {path: 'not-found', component: NotFoundComponent},
+  {path: '**', redirectTo: 'not-found', pathMatch: 'full'}
 ];
 
 @NgModule({