+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [RHa] - Version 1.1.0 - 2019-13-11
+### Initial setup
+* Add login component
+* Edit not-found component
+* Add bootstrap 
+* Add routes & redirects
+
 ## [RHa] - Version 1.0.0 - 2019-12-11
 ### Initial setup
 * Define basic project structure